requirements.txt
pandas==2.1.3
matplotlib==3.8.1
numpy==1.26.2
folium==0.15.1
geopy==2.4.1
flake8==4.0.1
pillow==9.0.1